+++ packages/components/components/elvis-breadcrumb/web_component.js
@@ -1,0 +1,1 @@
+export { default as Breadcrumb } from './dist/web_component/js/elvia-breadcrumb';