+++ packages/components/frameworks/react-ssr/src/App.js
@@ -2,11 +2,12 @@ import React, { useRef, useState } from 'react';
 import './App.scss';
 import { Popover } from '@elvia/elvis-popover/react';
 import { Accordion } from '@elvia/elvis-accordion/react';
-// import { Checkbox } from '@elvia/elvis-checkbox/react';
 import { Tabs } from '@elvia/elvis-tabs/react';
 import { ProgressLinear } from '@elvia/elvis-progress-linear/react';
 import { TestingComponent } from '@elvia/elvis-testing/react';
+import { Breadcrumb } from '@elvia/elvis-breadcrumb/react';
 import { Carousel } from '@elvia/elvis-carousel/react';
+import { Chip } from '@elvia/elvis-chip/react';
 import { Datepicker } from '@elvia/elvis-datepicker/react';
 import { Divider } from '@elvia/elvis-divider/react';
 import { Dropdown } from '@elvia/elvis-dropdown/react';
@@ -14,15 +15,30 @@ import { Box } from '@elvia/elvis-box/react';
 import { Modal } from '@elvia/elvis-modal/react';
 
 function App() {
-  // Old checkbox states
-  // const [trackedState, setTrackedState] = useState(true);
-  // const ref = useRef();
-  //////////////////////
+  const deletableChipsList = [
+    { value: 2022 },
+    { value: 2024, color: 'blue' },
+    { value: 2025, color: 'purple', disabled: true },
+  ];
+
+  const clickableChips = [
+    { value: 2022, color: 'green' },
+    { value: 2023, color: 'red' },
+    { value: 2024, color: 'blue', selected: true },
+    { value: 2025, color: 'purple' },
+    { value: 2026, color: 'violet', selected: true, disabled: true },
+  ];
+
+  const [chipSelected, setChipSelected] = useState(false);
+
   const [selectedState, setSelectedState] = useState(2);
+  const [chipsValues, setChipsValues] = useState([2018, 2019, 2020, 2021]);
+  const [deletableChips, setDeletableChips] = useState(deletableChipsList);
+  const [chipValue, setChipValue] = useState([]);
   const items = ['Statistikk', 'Siste kall', 'HAN-port', 'Feilkategorisering'];
   const dateCurr = new Date();
   const [isModalShowing, setIsModalShowingState] = useState(false);
-  const [isPopoverShowing, setIsPopoverShowingState] = useState(true);
+  const [isPopoverShowing, setIsPopoverShowingState] = useState(false);
   const [progressValue, setProgressValue] = useState(0);
 
   function increaseProgress() {
@@ -95,6 +111,21 @@ function App() {
     console.log(dropdownValue);
   };
 
+  const breadcrumbs = [
+    {
+      url: 'https://elvia.no',
+      title: 'Elvia.no',
+    },
+    {
+      url: 'https://www.elvia.no/nettleie',
+      title: 'Nettleie',
+    },
+    {
+      url: 'https://www.elvia.no/nettleie/elvias-leveringsplikt',
+      title: 'Elvias leveringsplikt',
+    },
+  ];
+
   const JSXCarouselElement = () => (
     <div>
       <p>
@@ -157,10 +188,36 @@ function App() {
     },
   ];
 
+  const handleOnDelete1 = (event) => {
+    const values = [...chipsValues];
+    setChipsValues(values.filter((value) => value !== event));
+  };
+
+  const handleOnDelete2 = (event) => {
+    const values = [...deletableChips];
+    setDeletableChips(values.filter((data) => data.value !== event));
+  };
+
+  const [filteredValues, setFilteredValues] = useState({
+    2022: false,
+    2023: false,
+    2024: true,
+    2025: false,
+    2026: true,
+  });
+
+  const handleOnValueChange = (event) => {
+    setFilteredValues((prevState) => ({
+      ...prevState,
+      [event.value]: event.isSelected,
+    }));
+  };
+
   return (
     <div className="App">
       <h1>React preview</h1>
-
+      <h2>Breadcrumbs</h2>
+      <Breadcrumb breadcrumbs={breadcrumbs} />
       <h2>Box</h2>
       <Box
         hasBorder={true}
@@ -168,7 +225,6 @@ function App() {
         title={<h1>Title for the box component</h1>}
         content={<div>Heisann dette er en box component sendt med som node i react</div>}
       ></Box>
-
       <h2>Modal</h2>
       <button onClick={() => setIsModalShowingState(true)}>Show modal</button>
       <Modal
@@ -263,7 +319,65 @@ function App() {
           </>
         }
       ></Modal>
-
+      <div>
+        <button
+          onClick={() => {
+            setChipsValues((values) => [...values, `20${Math.floor(Math.random() * 30) + 1}`]);
+          }}
+        >
+          Add chip
+        </button>
+      </div>
+      <div>
+        <h1>Test av chip</h1>
+        <button onClick={() => setChipSelected(!chipSelected)}>TEST</button>
+        <Chip
+          type="legend"
+          valueOnChange={(e) => {
+            setChipSelected(!chipSelected);
+          }}
+          selected={chipSelected}
+          value="Test"
+        ></Chip>
+        <hr></hr>
+      </div>
+      <div style={{ display: 'flex', flexDirection: 'row' }}>
+        {chipsValues.map((value) => (
+          <Chip value={value} onDelete={handleOnDelete1}></Chip>
+        ))}
+      </div>
+      Deletable Chips
+      <div style={{ display: 'flex', flexDirection: 'row' }}>
+        {deletableChips.map((data) => (
+          <Chip
+            value={data.value}
+            color={data.color}
+            disabled={data.disabled}
+            ariaLabel={`Fjern filtreringen for ${data.value}`}
+            onDelete={handleOnDelete2}
+          ></Chip>
+        ))}
+      </div>
+      Clickable Chips
+      <div style={{ display: 'flex', flexDirection: 'row' }}>
+        {clickableChips.map((data) => (
+          <Chip
+            value={data.value}
+            color={data.color}
+            selected={data.selected}
+            disabled={data.disabled}
+            type="legend"
+            valueOnChange={handleOnValueChange}
+          ></Chip>
+        ))}
+      </div>
+      Filters to apply: {`${Object.keys(filteredValues).filter((chip) => filteredValues[chip])}`}
+      <div style={{ margin: '40px 0' }}>
+        <h2>Clickable chip initally checkmark</h2>
+        <Chip value="Clickable3" type="choice" selected valueOnChange={setChipValue}></Chip>
+      </div>
+      <div>{'Selected chip: '}</div>
+      <div>{chipValue.isSelected ? chipValue.value : ''}</div>
       <h2>Dropdown</h2>
       <button onClick={consoleDropdownVal}>Console dropdown value</button>
       <div style={{ marginTop: '24px' }}>
@@ -298,7 +412,6 @@ function App() {
         <Carousel elements={elements.length} valueOnChange={setSelectedState} hideArrows></Carousel>
       </div>
       <div>{'Selected page: ' + selectedState.toString()}</div>
-
       <h2>Date picker</h2>
       <div
         style={{
@@ -319,11 +432,9 @@ function App() {
         <Datepicker isDisabled={true} valueOnChange={(dateCurr) => console.log(dateCurr)}></Datepicker>
         <Datepicker isCompact={true} valueOnChange={(dateCurr) => console.log(dateCurr)}></Datepicker>
       </div>
-
       <div style={{ margin: '40px 0' }}>
         <TestingComponent></TestingComponent>
       </div>
-
       <hr style={{ margin: '40px 0' }} />
       <h2>Divider</h2>
       <div style={{ marginBottom: '16px', padding: '8px' }}>
@@ -344,7 +455,6 @@ function App() {
       <div style={{ background: '#262626', padding: '8px' }}>
         <Divider type="curved" isInverted />
       </div>
-
       <h2>Accordion</h2>
       <div style={{ marginTop: '24px', marginBottom: '24px' }}>
         <Accordion
@@ -356,7 +466,6 @@ function App() {
           content="Bacon ipsum dolor amet pork loin bacon jowl turkey. Biltong sausage swine, shankle venison hamburger alcatra spare ribs bacon ham ribeye strip steak. Swine capicola picanha kevin drumstick. Chuck landjaeger pastrami, cow shoulder boudin short loin leberkas t-bone turkey prosciutto jowl. Turkey tail tongue cow shankle chicken tri-tip swine. Prosciutto pig ball tip kielbasa hamburger picanha pork chop tongue chicken shankle short loin filet mignon. T-bone shankle capicola, shoulder hamburger pancetta cupim chuck meatloaf turducken porchetta rump sausage strip steak ribeye."
         ></Accordion>
       </div>
-
       <h3>Custom Accordion content</h3>
       <div style={{ marginTop: '24px', marginBottom: '24px' }}>
         <Accordion
@@ -375,32 +484,13 @@ function App() {
           }
         ></Accordion>
       </div>
-
       <h2>Tabs</h2>
       <div style={{ marginTop: '16px', display: 'flex', justifyContent: 'center' }}>
         <Tabs items={items} value={2} valueOnChange={setSelectedState}></Tabs>
         <div>{selectedState.toString()}</div>
       </div>
-
       <hr style={{ margin: '40px 0' }} />
-
-      {/*       <h2>Checkbox</h2>
-      <div style={{ marginTop: '16px' }}>
-        <Checkbox
-          ref={ref}
-          checked={trackedState}
-          label="Normal checkbox"
-          name="Nametest"
-          id="CheckboxTestID"
-          size="normal"
-          changeHandler={setTrackedState}
-        ></Checkbox>
-        <Checkbox label="Small checkbox" name="Nametest" id="CheckboxTestID" size="small"></Checkbox>
-      </div>
- */}
-
       <hr style={{ margin: '40px 0' }} />
-
       <h2>Popover</h2>
       <div style={{ display: 'flex', flexDirection: 'column', alignItems: 'flex-end' }}>
         <Popover
@@ -498,25 +588,19 @@ function App() {
           posX="left"
         ></Popover>
       </div>
-
       <h2>Progressbar</h2>
-
       <ProgressLinear value={progressValue}></ProgressLinear>
-
       <div style={{ width: '100%', display: 'flex', justifyContent: 'space-around' }}>
         <button onClick={increaseProgress}>Increase</button>
         <button onClick={decreaseProgress}>Decrease</button>
         <button onClick={resetProgress}>reset</button>
       </div>
       <div>{progressValue}</div>
-
       <h2>Progressbar indeterminate</h2>
-
       <div>
         <ProgressLinear isIndeterminate></ProgressLinear>
       </div>
       <h2>Progressbar Error</h2>
-
       <div>
         <ProgressLinear isError></ProgressLinear>
       </div>