+++ packages/web/src/app/doc-pages/components/components.module.ts
@@ -24,9 +24,10 @@ import { AccordionDocModule } from './accordion-doc/accordion-doc.module';
 import { StepperDocModule } from './stepper-doc/stepper-doc.module';
 import { ContentLoaderDocModule } from './content-loader-doc/content-loader-doc.module';
 import { RadioFilterDocModule } from './radio-filter-doc/radio-filter-doc.module';
-import { ChipsModule } from './chips/chips.module';
+import { ChipDocModule } from './chip-doc/chip-doc.module';
 import { DividerDocModule } from './divider-doc/divider-doc.module';
 import { TabsDocModule } from './tabs-doc/tabs-doc.module';
+import { BreadcrumbDocModule } from './breadcrumb-doc/breadcrumb-doc.module';
 import { TimepickerDocModule } from './forms/timepicker-doc/timepicker-doc.module';
 import { BoxDocModule } from './box-doc/box-doc.module';
 
@@ -37,9 +38,10 @@ import { BoxDocModule } from './box-doc/box-doc.module';
     AutocompleteDocModule,
     BoxDocModule,
     ButtonDocModule,
+    BreadcrumbDocModule,
     CardDocModule,
     CarouselDocModule,
-    ChipsModule,
+    ChipDocModule,
     ComponentsStartModule,
     ContentLoaderDocModule,
     DividerDocModule,