+++ packages/web/src/app/doc-pages/dev/v2-playground/v2-playground.component.ts
@@ -17,7 +17,7 @@ export class v2PlaygroundComponent implements OnInit {
   isFullWidth = false;
   isModalShowing = false;
   date1 = new Date(2021, 4, 20);
-  isPopoverShowing = true;
+  isPopoverShowing = false;
 
   accordionContent = `
     Bacon ipsum dolor amet pork loin bacon jowl turkey. Biltong sausage swine, shankle venison hamburger alcatra spare ribs bacon ham ribeye strip steak. Swine capicola picanha kevin drumstick. Chuck landjaeger pastrami, cow shoulder boudin short loin leberkas t-bone turkey prosciutto jowl. Turkey tail tongue cow shankle chicken tri-tip swine. Prosciutto pig ball tip kielbasa hamburger picanha pork chop tongue chicken shankle short loin filet mignon. T-bone shankle capicola, shoulder hamburger pancetta cupim chuck meatloaf turducken porchetta rump sausage strip steak ribeye.
@@ -38,6 +38,18 @@ export class v2PlaygroundComponent implements OnInit {
   `;
 
   carouselValue = 0;
+
+  deleteValue = 0;
+
+  filteredValues = { 2021: false, 2022: true, 2023: true, 2024: true };
+  filteredKeys = Object.keys(this.filteredValues);
+
+  deletableChipsList = [
+    { value: 2022, color: 'green' },
+    { value: 2023, color: 'red' },
+    { value: 2024, color: 'blue' },
+    { value: 2025, color: 'purple', disabled: true },
+  ];
 
   elements = [
     {
@@ -99,6 +111,21 @@ export class v2PlaygroundComponent implements OnInit {
     },
   ];
 
+  breadcrumbsTest = [
+    {
+      url: 'https://elvia.no',
+      title: 'Elvia.no',
+    },
+    {
+      url: 'https://www.elvia.no/nettleie',
+      title: 'Nettleie',
+    },
+    {
+      url: 'https://www.elvia.no/nettleie/elvias-leveringsplikt',
+      title: 'Elvias leveringsplikt',
+    },
+  ];
+
   nativeJSApproach(): void {
     // Non-angular approach, not necessary when using angular:
     /*this.checkbox.nativeElement.addEventListener('checkedOnChange', (event: any) => {
@@ -112,21 +139,31 @@ export class v2PlaygroundComponent implements OnInit {
     this.nativeJSApproach();
   }
 
-  consoleLogDropdownValue = () => {
+  consoleLogDropdownValue = (): void => {
     console.log(this.dropdownValue);
   };
 
-  closeCallback = () => {
+  closeCallback = (): void => {
     console.log('Updated: closecallback');
     this.isModalShowing = !this.isModalShowing;
   };
 
-  updatePopoverVariable = (newValue: boolean) => {
+  updatePopoverVariable = (newValue: boolean): void => {
     console.log('Updated: closecallback ', newValue);
     this.isPopoverShowing = newValue;
   };
 
-  updatedVisibility = (event: any) => {
+  updatedVisibility = (event: any): void => {
     console.log('Updated - visibility: ' + event);
+  };
+
+  handleOnChange = (event: { value: string; isSelected: boolean }): void => {
+    this.filteredValues = { ...this.filteredValues, [event.value]: event.isSelected };
+  };
+
+  handleOnDelete = (event: number): void => {
+    this.deleteValue = event;
+    const values = [...this.deletableChipsList];
+    this.deletableChipsList = values.filter((value) => value.value !== event);
   };
 }