+++ packages/web/src/app/doc-pages/identity/icon-doc/icon-doc.component.ts
@@ -138,31 +138,6 @@ export class IconDocComponent implements OnInit {
     });
   }
 
-  toggleOpenDesktop(): void {
-    if (this.icons.nativeElement.classList.contains('open-accordion')) {
-      this.icons.nativeElement.classList.remove('open-accordion');
-    } else {
-      this.icons.nativeElement.classList.add('open-accordion');
-    }
-    if (this.accordionIconsDesktop.nativeElement.classList.contains('e-accordion__item--open')) {
-      this.accordionIconsDesktop.nativeElement.classList.remove('e-accordion__item--open');
-    } else {
-      this.accordionIconsDesktop.nativeElement.classList.add('e-accordion__item--open');
-    }
-  }
-  toggleOpenMobile(): void {
-    if (this.icons.nativeElement.classList.contains('open-accordion')) {
-      this.icons.nativeElement.classList.remove('open-accordion');
-    } else {
-      this.icons.nativeElement.classList.add('open-accordion');
-    }
-    if (this.accordionIconsMobile.nativeElement.classList.contains('e-accordion__item--open')) {
-      this.accordionIconsMobile.nativeElement.classList.remove('e-accordion__item--open');
-    } else {
-      this.accordionIconsMobile.nativeElement.classList.add('e-accordion__item--open');
-    }
-  }
-
   selectFilter(filter: string): void {
     this.selected = filter;
     if (filter === 'all') {