+++ packages/web/src/app/shared/component-documentation/component-example/component-example-generator/ceg-filters/ceg-filters.component.ts
@@ -11,11 +11,13 @@ export class CegFiltersComponent implements OnInit {
   @Input() componentData;
   @Input() props;
   @Input() selectedType;
-  codeWebComponentSub: Subscription;
+  codeAngularSub: Subscription;
   codeReactSub: Subscription;
+  codeNativeSub: Subscription;
   counterNumber: number;
   codeReact;
-  codeWebComponent;
+  codeAngular;
+  codeNative;
   checkboxesLength = 0;
   checkboxes = [];
   checkboxGroups = [];
@@ -27,19 +29,24 @@ export class CegFiltersComponent implements OnInit {
 
   ngOnInit(): void {
     this.codeReact = this.componentData.codeReact;
-    this.codeWebComponent = this.componentData.codeWebComponent;
-    this.codeWebComponentSub = this.cegService.listenCodeWebComponent().subscribe((code: string) => {
-      this.codeWebComponent = code;
+    this.codeAngular = this.componentData.codeAngular;
+    this.codeNative = this.componentData.codeNativeHTML;
+    this.codeAngularSub = this.cegService.listenCodeAngular().subscribe((code: string) => {
+      this.codeAngular = code;
     });
     this.codeReactSub = this.cegService.listenCodeReact().subscribe((code: string) => {
       this.codeReact = code;
+    });
+    this.codeNativeSub = this.cegService.listenCodeNative().subscribe((code: string) => {
+      this.codeNative = code;
     });
     this.initializeComponentProps();
   }
 
   ngOnDestroy(): void {
-    this.codeWebComponentSub.unsubscribe();
+    this.codeAngularSub.unsubscribe();
     this.codeReactSub.unsubscribe();
+    this.codeNativeSub.unsubscribe();
   }
 
   ngOnChanges(changes: SimpleChanges): void {
@@ -82,73 +89,52 @@ export class CegFiltersComponent implements OnInit {
     return checkboxArrays;
   }
 
-  updateProps(): void {
+  updateNewCode(): void {
     this.cegService.updateCodeReact(this.codeReact);
-    this.cegService.updateCodeWebComponent(this.codeWebComponent);
+    this.cegService.updateCodeAngular(this.codeAngular);
+    this.cegService.updateCodeNative(this.codeNative);
+  }
+
+  addNewProps(attr: string, newValue: string, type: string): void {
+    const elNameR = this.componentData.elementNameR;
+    const elNameW = this.componentData.elementNameW;
+    this.codeReact = this.cegService.addNewProp(this.codeReact, attr, newValue, 'react', type, elNameR);
+    this.codeAngular = this.cegService.addNewProp(this.codeAngular, attr, newValue, 'angular', type, elNameW);
+    this.codeNative = this.cegService.addNewProp(this.codeNative, attr, newValue, 'native', type, elNameW);
+  }
+
+  replaceOldProps(attr: string, newValue: string, type: string): void {
+    this.codeReact = this.cegService.replaceOldProp(this.codeReact, attr, newValue, 'react', type);
+    this.codeAngular = this.cegService.replaceOldProp(this.codeAngular, attr, newValue, 'angular', type);
+    this.codeNative = this.cegService.replaceOldProp(this.codeNative, attr, newValue, 'native', type);
+  }
+
+  removeProps(attr: string): void {
+    this.codeReact = this.cegService.removeProp(this.codeReact, attr);
+    this.codeAngular = this.cegService.removeProp(this.codeAngular, attr);
+    this.codeNative = this.cegService.removeProp(this.codeNative, attr);
   }
 
   updateRadioProp(prop: any, newValue: string): void {
-    if (this.codeWebComponent.includes(prop.attribute)) {
-      // Replaces old value for prop
-      this.codeReact = this.codeReact.replace(
-        this.cegService.getPropRegex(prop.attribute),
-        this.cegService.getReplaceValueString(prop.attribute, newValue, true, prop.cegType),
-      );
-      this.codeWebComponent = this.codeWebComponent.replace(
-        this.cegService.getPropRegex(prop.attribute),
-        this.cegService.getReplaceValueString(prop.attribute, newValue, false, prop.cegType),
-      );
+    const attr = prop.attribute;
+    const type = prop.cegType;
+    if (this.codeAngular.includes(prop.attribute)) {
+      this.replaceOldProps(attr, newValue, type);
     } else {
-      // Adds new prop to code
-      const newLineRegexW = this.cegService.getNewLineRegex(this.componentData.elementNameW);
-      const newStringW = this.cegService.getNewPropStringW(
-        this.componentData.elementNameW,
-        prop.attribute,
-        newValue,
-      );
-      this.codeWebComponent = this.codeWebComponent.replace(newLineRegexW, newStringW);
-
-      const newLineRegexR = this.cegService.getNewLineRegex(this.componentData.elementNameR);
-      const newStringR = this.cegService.getNewPropStringR(
-        this.componentData.elementNameR,
-        prop.attribute,
-        newValue,
-        prop.cegType,
-      );
-      this.codeReact = this.codeReact.replace(newLineRegexR, newStringR);
+      this.addNewProps(attr, newValue, type);
     }
-    this.updateProps();
+    this.updateNewCode();
   }
 
-  updateToggleCheckboxProp(prop: any): void {
-    if (this.codeWebComponent.includes(prop.attribute)) {
-      // Removes old prop and line in code
-      this.codeReact = this.codeReact
-        .replace(this.cegService.getPropRegex(prop.attribute), '')
-        .replace(this.emptyLineRegex, '');
-      this.codeWebComponent = this.codeWebComponent
-        .replace(this.cegService.getPropRegex(prop.attribute), '')
-        .replace(this.emptyLineRegex, '');
+  updateToggleCheckboxProp(prop: any, newValue: string): void {
+    const attr = prop.attribute;
+    const type = prop.cegType;
+    if (this.codeAngular.includes(prop.attribute)) {
+      this.removeProps(attr);
     } else {
-      // Adds new prop in code
-      const newLineRegexW = this.cegService.getNewLineRegex(this.componentData.elementNameW);
-      const newStringW = this.cegService.getNewPropStringW(
-        this.componentData.elementNameW,
-        prop.attribute,
-        prop.cegOption,
-      );
-      this.codeWebComponent = this.codeWebComponent.replace(newLineRegexW, newStringW);
-
-      const newLineRegexR = this.cegService.getNewLineRegex(this.componentData.elementNameR);
-      const newStringR = this.cegService.getNewPropStringR(
-        this.componentData.elementNameR,
-        prop.attribute,
-        prop.cegOption,
-        prop.cegType,
-      );
-      this.codeReact = this.codeReact.replace(newLineRegexR, newStringR);
+      this.addNewProps(attr, newValue, type);
     }
-    this.updateProps();
+    this.updateNewCode();
   }
 
   isAcceptedCounterValue(prop: any, newValue: number): boolean {
@@ -160,6 +146,8 @@ export class CegFiltersComponent implements OnInit {
   }
 
   updateCounterProp(prop: any, newValue: number): void {
+    const attr = prop.attribute;
+    const type = prop.cegType;
     if (this.isAcceptedCounterValue(prop, newValue)) {
       return;
     } else if (this.counterNumber === undefined) {
@@ -168,27 +156,10 @@ export class CegFiltersComponent implements OnInit {
       this.counterNumber += newValue;
     }
 
-    if (this.codeWebComponent.includes(prop.attribute)) {
+    if (this.codeAngular.includes(attr)) {
       // Replaces old value for prop
-      this.codeReact = this.codeReact.replace(
-        this.cegService.getPropRegex(prop.attribute),
-        this.cegService.getReplaceValueString(
-          prop.attribute,
-          this.counterNumber.toString(),
-          true,
-          prop.cegType,
-        ),
-      );
-      this.codeWebComponent = this.codeWebComponent.replace(
-        this.cegService.getPropRegex(prop.attribute),
-        this.cegService.getReplaceValueString(
-          prop.attribute,
-          this.counterNumber.toString(),
-          false,
-          prop.cegType,
-        ),
-      );
+      this.replaceOldProps(attr, '' + this.counterNumber, type);
     }
-    this.updateProps();
+    this.updateNewCode();
   }
 }