+++ packages/components/components/elvis-dropdown/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Elvia Dropdown
 
+## 1.0.6 (09.09.21)
+
+### Bug fixes
+
+Fix for width-bug inside a flex-container.
+
+## 1.0.5 (10.08.21)
+
+### Bug fixes
+
+Updated styled component version to fix warning about different versions.
+
 ## 1.0.4 (08.06.21)
 
 ### Bug fixes